Major update to controller, robot object, kinematic solver, and models
This is a major update to the codebase. After a month of testing out different kinematic solvers and investigating other Github repos working with robot arms, I've settled with the current configuration. It utilizes the advantage of having a small enough footprint to fit the code and supporting libraries on the PICO while making quick calculations with the kinematic solver. I also added another joint to the robot in order to make it a true 6DOF robot arm. 

Major revisions:
 - Added the `3d_files` subfolder to hold the actual models for the robot
 - Added the `test` subfolder to document the different libraries tested
 - Created a new `AsyncController` controller object which utilizes the `asyncio` library and performs multiple coroutines:
    - It can connect to a network and wait for messaged to be received from a client independently  from its operation, and just stores new messaged in a queue-like buffer
    - The internal controller state will parse messages that are dequeuing and handle them
    - It handles multiple routines similar to threads but uses less memory and avoids race conditions
 - Updated the `RobotArm` object to  how handle forward and inverse kinematics (yay!). Benchtesting shows the update rate for new pose calculations to be up to 50Hz
 - Added a `XBoxClient` package for some early xbox controller communication (still needs tuning)

Minor updates:
 - `music_themes.py` has been replaced with `musicplayer.py` 
 - Updates to the README file to include the full command suite supported
 - `settings.toml` has the default network credentials (don't want mine in there!)

Next should come the physical testing with the actual servos, and if all goes well the videos will follow!