Inverse Kinematic and Trajectories now supported!
Major announcement: Inverse Kinematics can now be handled! YAY!

The `robotarm` module now has a `Robot` class with the preconfigured forward kinematic transform from the base to the gripper, given the physical parameters of the arm and the axes of rotations for each joint. The IK needed to be handled with an analytical solver instead of a numerical solver due to the limited size of the PICO (2MB, when formatted, 500KB when flashing the CircuitPython image), and the ulab library which contains the MicroPython equivalent of `numpy` has only barebones functionality, so no `numpy.scipy.optimize` or symbolic solvers 🤔 hmm... 

The controller now comes from the `async_controller` module where the `AsyncController` object handles the communication between connected clients and commands parsed to the robot. Fastest update time with handling a circular trajectory with IK pose commands (with an array) was 30Hz, but the queueing/dequeueing methods which handle some extra 'utf-8' decoding and command parsing drop that down to 12Hz. Might be faster with a usb connection or a single command with a list of points to hit, but it's great that it run's sort of smoothly with controller inputs!

Also pushing a few more test files used to make sure the asynchronous multi-servo commands would work. 

A follow-up update should include touch-ups with the documentation, instructions, code comments, gifs, and media posts.