Major updates to robot firmware
Major updates:
1) Robot models have been updated, created in Solidworks but STL files should be the same
2) Servo models downloaded for the complete assembly
3) urdf created in `urdf_export` package, will be necessary for ROS2 drivers, need to do some organization later
4) New option to set whether to use console direct communication and the `usb_cdc` package, or by using the UART pins for TX/RX communication from the builtin `board` package.

Minor changes:
1) Comparison files using numerical and symbolic methods to compare versions.
2) Highlighted difference between using regular DH parameters and DH modified parameters. Building the homogenous transformation matrix using one method isn't the same as the other
3) Polished the robot models for the miniarm and KUKA test model
4) Button command for the xbox USB serial control

A lot of progress was made this week using the custom AJ GUI designed in Unity, and using the [Ardity](https://ardity.dwilches.com/) package. A discovery was made in which the Arduino had no issues with communicating because the USB connection port supported UART communication by default, but the Raspberry Pi Pico W microcontroller did not with its usb connection port. We set 2 of the GPIO pins for UART communication and used a USB-to-UART serial converter board to send commands from the Unity GUI to the robot. A separate microusb cable was plugged into the Pico with the other end to a laptop, and PUTTY could be used to observe the environment and listen to the robot code while serial commands were being sent from the GUI! Hard to figure out several issues with reading and writing at first, but very Impressive that it worked in the end!

Next steps from here are:
1)  figuring out the `set_joint_speed_control()` function that updates all the robot joint values but takes into account steps and rotation velocities.
2) Verify delta commands moving away from home and returning to home will compute the correct joint angles.